moving get_market_data into management folder per django docs
